Updated variables and constructors in Employee class

+++ JavaWebService/src/main/java/com/kainos/ea/Payable.java
@@ -1,7 +1,8 @@
 package com.kainos.ea;
 
 public interface Payable {
-    public int calcPay();
+    //Changed from int to float to match the changes in Employee class
+    public float calcPay();
 
     public String getName();
     public String setName();
